% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/probabilities_compared_to_previously_published.R
\name{add_conditional_formatting_to_trends}
\alias{add_conditional_formatting_to_trends}
\title{add_conditional_formatting_to_trends}
\usage{
add_conditional_formatting_to_trends(probs_over_time_prep)
}
\arguments{
\item{probs_over_time_prep}{outputted probs_over time}
}
\value{
data.frame
}
\description{
adds conditional formatting to probability/trend tables
}
