% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_region_or_country_layer.R
\name{add_region_or_country_layer}
\alias{add_region_or_country_layer}
\title{Add region facet to plots}
\usage{
add_region_or_country_layer(plot, region_or_country)
}
\arguments{
\item{plot}{plot already made}

\item{region_or_country}{specify "country" for one ggplot object of data from an individial country, specify "region" for ggplot object of England regions}
}
\value{
a ggplot object
}
\description{
Add facets of regions when specified by region_or_country
}
