% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{add_trend_summary}
\alias{add_trend_summary}
\title{Add Trend Summary}
\usage{
add_trend_summary(likelihood_data, country_configs, country)
}
\arguments{
\item{likelihood_data}{a dataframe containing the probability scores}

\item{country_configs}{a list object containing the configuration settings by country}

\item{country}{a string containing the name of the country you want to filter by}
}
\value{
a data frame
}
\description{
function to determine the trend_statement from the probability scores
}
