% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{aggregate_england_from_regions_and_date}
\alias{aggregate_england_from_regions_and_date}
\title{Aggregate England by Region and Time Columns}
\usage{
aggregate_england_from_regions_and_date(data)
}
\arguments{
\item{data}{a data frame containing two columns (region and time)}
}
\value{
a dataframe with England as a region, as well as the NUTS1 regions
}
\description{
function to create England from the NUTS1 England regions
}
