% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{aggregate_raw_england_region_stats}
\alias{aggregate_raw_england_region_stats}
\title{aggregate_raw_england_region_stats}
\usage{
aggregate_raw_england_region_stats(data, country_filter, region_or_country)
}
\arguments{
\item{data}{a data frame from which you want to caluclate the statistics}

\item{country_filter}{a string containing the country to be filtered}

\item{region_or_country}{a string containing either 'region' or 'country'}
}
\value{
a dataframe with aggregated region statistics
}
\description{
function to aggregate England region raw statistics for comparison with modelled data
}
