% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_external_libraries.R
\name{attach_external_libraries}
\alias{attach_external_libraries}
\title{attach_external_libraries}
\usage{
attach_external_libraries()
}
\value{
None
}
\description{
This function attaches magittr, ggplot2 and patchwork to the namespace
}
