% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prevalence_series.R
\name{calculate_means_and_cis}
\alias{calculate_means_and_cis}
\title{calculate_means_and_cis}
\usage{
calculate_means_and_cis(draws)
}
\arguments{
\item{draws}{a dataframe containing the poststratified draws for a specific region, country and variant, with each column containing draws for each date.}
}
\value{
a dataframe containing the date, mean, ll (lower credible interval), and ul (upper credible interval)
}
\description{
function to calculate means and credible intervals for the set of draws you have provided.
}
