% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{calculate_probabilities_comparing_dates}
\alias{calculate_probabilities_comparing_dates}
\title{calculate_probabilities_comparing_dates}
\usage{
calculate_probabilities_comparing_dates(
  post_stratified_draws,
  country_configs,
  region_list,
  country_list,
  variant_list
)
}
\arguments{
\item{post_stratified_draws}{a list object containing the post-stratified draws from the model}

\item{country_configs}{a list object containing the configuration settings for each country}

\item{region_list}{a string containing the name of the region you want to filter by}

\item{country_list}{a string containing the name of the country you want to filter by}

\item{variant_list}{a string containing the name of the variant you want to filter by}
}
\value{
a matrix containing the predicted draws for the specific region, country and variant
}
\description{
a function that iterates ofver the calcualte probabilities_by_region_country_and_variant_function and binds the outputs together into a single dataframe.
}
