% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_data_run_vs_previous}
\alias{check_data_run_vs_previous}
\title{check_data_run_vs_previous}
\usage{
check_data_run_vs_previous(config)
}
\arguments{
\item{config}{- a list object containing the configurations for the system}
}
\value{
error if datarun does not occur after the previous data run
}
\description{
checking data run occurs after the previous data run
}
