% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_dual_parameters_are_not_both_true}
\alias{check_dual_parameters_are_not_both_true}
\title{check_dual_parameters_are_not_both_true}
\usage{
check_dual_parameters_are_not_both_true(config)
}
\arguments{
\item{config}{a list object containing the configurations settings}
}
\value{
error if the test and rerun are set to true while test config short is set to true in the config
}
\description{
checking test and rerun are not set to true while test config short is set to true in the config
}
