% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_dual_parameters_are_not_both_true_by_param}
\alias{check_dual_parameters_are_not_both_true_by_param}
\title{check_dual_parameters_are_not_both_true_by_param}
\usage{
check_dual_parameters_are_not_both_true_by_param(param_1, param_2, config)
}
\arguments{
\item{param_1}{- a string containing the first reference to test and rerun in the config}

\item{param_2}{- a string containing the second reference to test_configs_short}

\item{config}{- a list object containing the configurations for the system}
}
\value{
error if the test and rerun are set to true while test config short is set to true in the config
}
\description{
checking test and rerun are not set to true while test config short is set to true in the config
}
