% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_end_date_occurs_after_data_run}
\alias{check_end_date_occurs_after_data_run}
\title{check_end_date_occurs_after_data_run}
\usage{
check_end_date_occurs_after_data_run(config)
}
\arguments{
\item{config}{- a list object containing the configurations for the system}
}
\value{
error if any of the countries end dates occur after the data runs
}
\description{
checking end dates occurs after the data runs
}
