% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_end_date_occurs_after_data_run_by_vars}
\alias{check_end_date_occurs_after_data_run_by_vars}
\title{check_end_date_occurs_after_data_run_by_vars}
\usage{
check_end_date_occurs_after_data_run_by_vars(
  country,
  data_run,
  end_date,
  config
)
}
\arguments{
\item{country}{- a string containing the name of a country in the dataset}

\item{data_run}{- a string containing the reference to a data run variable in the config}

\item{end_date}{- a string containing the reference to an end date variable in the config}

\item{config}{- a list object containing the configurations for the system}
}
\value{
error if the country end dates occur after the data runs
}
\description{
checking end dates occurs before the data runs by country
}
