% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{check_end_date_occurs_after_prev_end_date_by_vars}
\alias{check_end_date_occurs_after_prev_end_date_by_vars}
\title{check_end_date_occurs_after_prev_end_date_by_vars}
\usage{
check_end_date_occurs_after_prev_end_date_by_vars(country, config)
}
\arguments{
\item{country}{- a string containing the name of a country in the dataset,}

\item{config}{- a list object containing the configurations for the system}
}
\value{
error if the country end date occurs before the previous end date
}
\description{
checking end date occurs after the previous end date by country
}
