% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_dates.R
\name{check_missing_date}
\alias{check_missing_date}
\title{check_missing_date}
\usage{
check_missing_date(data, country, check_date, date_name)
}
\arguments{
\item{data}{a dataframe which you want to check the data (must have visit_date column)}

\item{country}{a string of the name of the country on which you want to run this on}

\item{check_date}{a date value for the check}

\item{date_name}{a string containing the name of the data (e.g. 'start_date')}
}
\description{
return an error if a specific date is missing
}
