% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_dates.R
\name{check_missing_dates}
\alias{check_missing_dates}
\title{check missing dates in the aggregated data}
\usage{
check_missing_dates(data, config)
}
\arguments{
\item{data}{the aggregated sample data}

\item{config}{the main config list}
}
\description{
check all dates needed for the analysis are present in the sample for all countries.
}
