% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_dates.R
\name{check_missing_dates_between}
\alias{check_missing_dates_between}
\title{check_missing_dates_between}
\usage{
check_missing_dates_between(data, country, start_date, end_date)
}
\arguments{
\item{data}{a dataframe which you want to check the data (must have visit_date column)}

\item{country}{a string of the name of the country on which you want to run this on}

\item{start_date}{the start date for the check}

\item{end_date}{the end date for the check}
}
\description{
return a warning if there are any missing visit_dates between start_date and end_date
}
