% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_dates.R
\name{check_missing_dates_by_country}
\alias{check_missing_dates_by_country}
\title{check missing dates in the aggregated data by country}
\usage{
check_missing_dates_by_country(country_select, data, config)
}
\arguments{
\item{country_select}{the geography to check}

\item{data}{the aggregated sample data}

\item{config}{the main config list}
}
\description{
check all dates needed for the analysis are present in the sample by country. Stops the script if the last day in the series is missing, as this will break the model later on.
}
