% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_aggregated.R
\name{clean_aggregated}
\alias{clean_aggregated}
\title{clean aggregated sample}
\usage{
clean_aggregated(data)
}
\arguments{
\item{data}{aggregated sample dataframe}
}
\description{
basic data cleaning
}
