% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_population_counts.R
\name{clean_population_counts}
\alias{clean_population_counts}
\title{clean_population_counts}
\usage{
clean_population_counts(population_counts)
}
\arguments{
\item{population_counts}{a dataframe with the aggregated population counts by demographic}
}
\value{
dataframe
}
\description{
function to reformat the population counts
}
