% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_region.R
\name{clean_region}
\alias{clean_region}
\title{clean_region}
\usage{
clean_region(data, column)
}
\arguments{
\item{data}{a dataframe containing the region codes for renaming}

\item{column}{a column name which contains the region codes for renaming}
}
\value{
a dataframe with a readable region column which is of the 'factor' type
}
\description{
function recode and reformat the region identifier column to readable names within a factor type
}
