% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_overlay_plot.R
\name{combine_current_and_previous_prevalence}
\alias{combine_current_and_previous_prevalence}
\title{Combine Current and Previous Prevalence}
\usage{
combine_current_and_previous_prevalence(current, previous)
}
\arguments{
\item{current}{a dataframe containing the prevalence time series}

\item{previous}{a dataframe containing the previous prevalence time series}
}
\value{
a data frame
}
\description{
binds current and previous prevalence data frames together
}
