% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{combine_draws_to_single_df}
\alias{combine_draws_to_single_df}
\title{Combine Draws to a Single Dataframe}
\usage{
combine_draws_to_single_df(post_stratified_draws, main_config)
}
\arguments{
\item{post_stratified_draws}{a list object containing the post stratified draws matricies}

\item{main_config}{a list object containing the main configuration settings for the system}
}
\value{
transformed_draws a dataframe with all of the draws broken down by country, region, variant and date
}
\description{
combines post_stratified_draws from a list of matricies to a single data frame
}
