% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/probabilities_compared_to_previously_published.R
\name{combine_probabilities}
\alias{combine_probabilities}
\title{combine_probabilities}
\usage{
combine_probabilities(probabilities, previous_probabilities)
}
\arguments{
\item{probabilities}{a dataframe containing the current probabilities}

\item{previous_probabilities}{a dataframe containing the previous probabitlities}
}
\value{
a dataframe with the combined probablities (both current and previous)
}
\description{
function to bind together current and previous probabilities into one dataframe
}
