% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coloured_trends_table.R
\name{create_coloured_trends_table}
\alias{create_coloured_trends_table}
\title{Create Coloured Trends Table}
\usage{
create_coloured_trends_table(data, COUNTRY, VARIANT, region_or_country)
}
\arguments{
\item{data}{a dataframe containing the post-stratified probabilities with formatting}

\item{COUNTRY}{a string containing the name of the region you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}

\item{region_or_country}{a string containing either region or country}
}
\value{
a kable formatted table for rendering in HTML
}
\description{
makes sequence of dates from start to end date
}
