% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{create_comparison_dates_dataframe}
\alias{create_comparison_dates_dataframe}
\title{create_comparison_dates_dataframe}
\usage{
create_comparison_dates_dataframe(draw, country_configs, country)
}
\arguments{
\item{draw}{a matrix of draws (rows) by date (cols)}

\item{country_configs}{a list object containing the country configuration settings}

\item{country}{a string containing the name of the country you want to filter by}
}
\value{
a dataframe of the reference day and comparison day draws
}
\description{
function to create a dataframe of reference day and comparison date draws
}
