% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_aggregated.R
\name{create_country_from_region}
\alias{create_country_from_region}
\title{create_country_from_region}
\usage{
create_country_from_region(data)
}
\arguments{
\item{data}{a dataframe containing a region column}
}
\value{
a dataframe with an additional country column
}
\description{
function to infer country name from region column
}
