% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{create_file_name}
\alias{create_file_name}
\title{create_file_name}
\usage{
create_file_name(
  file_id,
  end_date,
  file_type = c("rdata", "yaml", "jpeg", "html", "csv", "xlsx", "rds")
)
}
\arguments{
\item{file_id}{a string containing the identifying characteristics of the object}

\item{end_date}{a cleaned string containing the date in the following format yyyymmdd}

\item{file_type}{a list containing strings the file extension without the '.' e.g. 'rdata' or 'csv'}
}
\value{
a string containing the file_name
}
\description{
function to create a file_name for saving outputs
}
