% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{create_filepath}
\alias{create_filepath}
\title{create_filepath}
\usage{
create_filepath(
  file_id,
  end_date,
  file_type = c("rdata", "yaml", "jpeg", "html", "csv", "xlsx", "rds"),
  config
)
}
\arguments{
\item{file_id}{a string containing the identifying characteristics of the object}

\item{end_date}{a cleaned string containing the date in the following format yyyymmdd}

\item{file_type}{a list containing strings the file extension without the '.' e.g. 'rdata' or 'csv'}

\item{config}{a list object containing the configurations settings}
}
\value{
a string containing the filepath
}
\description{
function to create a filepath for saving outputs
}
