% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{create_folder_name}
\alias{create_folder_name}
\title{create_folder_name}
\usage{
create_folder_name(config)
}
\arguments{
\item{config}{a list object containing the main configuration settings for the system}
}
\value{
a string containing the folder name
}
\description{
function to create folder name from config settings
}
