% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prevalence_series.R
\name{create_prevalence_by_region_country_and_variant}
\alias{create_prevalence_by_region_country_and_variant}
\title{create_prevalence_by_region_country_and_variant}
\usage{
create_prevalence_by_region_country_and_variant(
  region,
  country,
  variant,
  post_stratified_draws,
  country_configs
)
}
\arguments{
\item{region}{a string containing the name of the region you want to filter by}

\item{country}{a string containing the name of the country you want to filter by}

\item{variant}{a string containing the name of the variant you want to filter by}

\item{post_stratified_draws}{a list of poststratified draws (country x variant)}

\item{country_configs}{a list object containing the country configuration settings}
}
\value{
a dataframe with summarised prevalence estimates by region, country, variant
}
\description{
function to create summarised prevalence estimates (mean, ll, ul) by region country and variant
}
