% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prevalence_series.R
\name{create_prevalence_series}
\alias{create_prevalence_series}
\title{create_prevalence_series}
\usage{
create_prevalence_series(
  country_configs,
  post_stratified_draws,
  region_list,
  country_list,
  variant_list
)
}
\arguments{
\item{country_configs}{a list of country configs}

\item{post_stratified_draws}{a list of poststratified draws (country x variant)}

\item{region_list}{a list of the regions}

\item{country_list}{A list of the four UK countries, England, Scotland, Wales and Northern_Ireland}

\item{variant_list}{a list of the variant 1, 2, 4, 5}
}
\description{
gets the prevalence time series output from the list of poststratified draws
}
