% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{create_raw_counts_with_prevalence}
\alias{create_raw_counts_with_prevalence}
\title{Create Raw Counts with Prevalence}
\usage{
create_raw_counts_with_prevalence(
  aggregated_sample,
  aggregated_HH,
  prevalence_time_series
)
}
\arguments{
\item{aggregated_sample}{a dataframe containing the raw participant sample}

\item{aggregated_HH}{a dataframe containing the raw household sample}

\item{prevalence_time_series}{a dataframe containing the modelled prevalence estimates}
}
\value{
a dataframe with the modelled estimates combined with raw counts
}
\description{
a function to join the cleaned raw data with the modelled prevalence estimates
}
