% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{create_region_variant_country_combos}
\alias{create_region_variant_country_combos}
\title{Create Region, Variant, and Country Combinations}
\usage{
create_region_variant_country_combos(main_config)
}
\arguments{
\item{main_config}{a list object containing the main configuration settings for the system}
}
\value{
a dataframe with all of the combinations of region, variant and country that occur in the post_stratified_draws
}
\description{
create a dataframe with all of the combinations of Region, Variant and Country which exist within the post_stratified_draws matrix
}
