% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aggregated_samp.R
\name{create_result_by_variant}
\alias{create_result_by_variant}
\title{create result by variant}
\usage{
create_result_by_variant(data, variant_type)
}
\arguments{
\item{data}{a dataframe containing variant_type, n_pos, n_ctpattern[4,5,6,7]}

\item{variant_type}{integer should be 1, 2, 4 or 5}
}
\value{
data frame
}
\description{
create result column by variant type
}
