% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variant_overlay.R
\name{create_variant_overlay}
\alias{create_variant_overlay}
\title{Create Variant Overlay}
\usage{
create_variant_overlay(
  prevalence_time_series,
  COUNTRY,
  main_config,
  region_or_country
)
}
\arguments{
\item{prevalence_time_series}{MRP modelled_data series}

\item{COUNTRY}{character}

\item{main_config}{a list object containing the configuration settings for the system}

\item{region_or_country}{character either "region" or "country"}
}
\value{
plot
}
\description{
creates a plot comparing each variant over time
}
