% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_clean_string.R
\name{date_to_clean_string}
\alias{date_to_clean_string}
\title{date_to_clean_string}
\usage{
date_to_clean_string(date)
}
\arguments{
\item{date}{a date type object <yyyy-mm-dd>}
}
\value{
reformatted date
}
\description{
convert date type object to yyyymmdd string
}
