% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_config}
\alias{dummy_config}
\title{Dummy Config}
\format{
An object of class \code{list} of length 4.
}
\source{
created by data-raw/make_dummy_config.R
}
\usage{
dummy_config
}
\description{
A dummy configuration file to run unit tests on automated file disclosure control checking functions.
}
\keyword{datasets}
