% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_country_configs}
\alias{dummy_country_configs}
\title{Dummy Country Configs}
\format{
An object of class \code{list} of length 4.
}
\source{
created by data-raw/make_dummy_country_configs.R
}
\usage{
dummy_country_configs
}
\description{
A dummy country specific configuration file to run unit tests on automated file disclosure control checking functions.
}
\keyword{datasets}
