% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_post_stratified_draws}
\alias{dummy_post_stratified_draws}
\title{Dummy Post Stratified Draws}
\format{
An object of class \code{list} of length 8.
}
\source{
created by data-raw/make_dummy_post_stratified_draws.R
}
\usage{
dummy_post_stratified_draws
}
\description{
A dummy dataset used for examples and/or testing. Represents what post-stratified draws should look like after they have been produced. This dummy data represents the England and Scotland data.
}
\keyword{datasets}
