% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_reference_day_vs_weeks_before_probabilities}
\alias{dummy_reference_day_vs_weeks_before_probabilities}
\title{Dummy Reference Day vs Weeks before Probabilities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 8 columns.
}
\source{
created by data-raw/make_dummy_post_stratified_draws.R
}
\usage{
dummy_reference_day_vs_weeks_before_probabilities
}
\description{
A dummy dataset used for examples and/or testing. Represents what post-stratified Reference Day vs Weeks before Probabilities should look like after they have been produced. This dummy data represents the England and Scotland data.
}
\keyword{datasets}
