% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pop_table.R
\name{duplicate_pop_table}
\alias{duplicate_pop_table}
\title{Duplicate population tables}
\usage{
duplicate_pop_table(pop_table, config)
}
\arguments{
\item{pop_table}{data frame}

\item{config}{a config list containing $run_settings$n_days_to_model}
}
\value{
data frame
}
\description{
Duplicate input data and add it to the bottom of the data.
repeat duplication of original input data a specified number of times.
}
