% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_probabilities.R
\name{extract_posterior_probabilities}
\alias{extract_posterior_probabilities}
\title{Extract probabilities from an MRP model, per region if needed}
\usage{
extract_posterior_probabilities(model, post_strat_table, aggregated_data)
}
\arguments{
\item{model}{GAMM4 model fit object}

\item{post_strat_table}{post_strat_table, replicated as many times as there
are days in the model. n rows should equal the full factorial product of
the n levels in all the factors in the model design n levels region x
n levels var 2, x n levels var 3 etc}

\item{aggregated_data}{aggregated sample data}
}
\value{
list of matrices (numeric) where one matrix contains all the posterior
        probabilities (i.e. for the whole country) and the other matrices contain
        the posterior probabilities for each of the level of the region variable
        in the model:
        - rows =  n draws sampled during modelling
        - for the averge of a factor, n col = the full factorial design product
          (i.e. n levels region x n levels var 2, x n levels var 3 etc - this equates
          to the n rows in the poststrat table).
        - for the single factor level matrices, n columns is equal to the
          product of number of levels in all factors besides the region variable
          (i.e. n levels var 2, x n levels var 3 etc)
}
\description{
applies rstanarm::posterior_epred
             and produces a list of  posterior draws of the predictor.
             IMPORTANT - for this to work, the workspace must contain two variables
             containing the oucome variable. this is an inssue with the
             rstanarm ::posterior_epred function rather than this function.
             these two necessary variables can be extracted from the model
             object like so:
             - n_pos <- model$model$`cbind(n_pos, n_neg)`[,1]
}
