% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aggregated_samp.R
\name{filter_aggregated_samp}
\alias{filter_aggregated_samp}
\title{filter_aggregated_sample}
\usage{
filter_aggregated_samp(
  aggregated_sample,
  country_configs,
  country_list,
  variant_list
)
}
\arguments{
\item{aggregated_sample}{cleaned aggregated_sample to be filtered for the country x variant models to be fitted}

\item{country_configs}{a list of country configs}

\item{country_list}{A list of the four UK countries, England, Scotland, Wales and Northern_Ireland}

\item{variant_list}{a list of the variant 1,2,4,5}
}
\description{
filters and cleans the whole uk aggregated sample into DAs and England samples
}
