% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_country.R
\name{filter_by_country}
\alias{filter_by_country}
\title{filter by country}
\usage{
filter_by_country(data, country)
}
\arguments{
\item{data}{data frame}

\item{country}{country to keep in the data ("England", "Wales", "Scotland", "NI")}
}
\value{
data frame
}
\description{
fiter data by country
}
