% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{filter_by_region_or_country}
\alias{filter_by_region_or_country}
\title{filter_by_region_or_country}
\usage{
filter_by_region_or_country(data, country_filter, region_or_country)
}
\arguments{
\item{data}{a data frame containing the dataset to filter}

\item{country_filter}{string continaing the name of a specific country for which to filter}

\item{region_or_country}{a string containing either 'region' or 'country'}
}
\value{
dataframe which has been filtered keeping only specific regions
}
\description{
filters dataframe and keeps specific regions or countries
}
