% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{filter_dates}
\alias{filter_dates}
\title{filter dates}
\usage{
filter_dates(data, country_configs, country)
}
\arguments{
\item{data}{data frame}

\item{country_configs}{list object containing the countries config}

\item{country}{a string containing the name of the country you want to filter by}
}
\value{
data frame
}
\description{
filter dates outside of the date range for the analysis
}
