% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aggregated_samp.R
\name{filter_days_with_no_returns}
\alias{filter_days_with_no_returns}
\title{filter days with no returns}
\usage{
filter_days_with_no_returns(data)
}
\arguments{
\item{data}{a dataframe which could contain days with no returns}
}
\description{
filter test results to remove those with no returns
}
