% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_region_country_and_variant.R
\name{filter_for_region_country_and_variant}
\alias{filter_for_region_country_and_variant}
\title{filter_for_region_country_and_variant}
\usage{
filter_for_region_country_and_variant(
  regions_filter,
  countries_filter,
  variants_filter
)
}
\arguments{
\item{regions_filter}{a vector containing the names of the regions to filter using "All" for country level region id's at this stage}

\item{countries_filter}{a vector containing the names of the countries to filter}

\item{variants_filter}{a vector containing the names of the variants to filter}
}
\value{
a dataframe which has been filtered by region, country and variant
}
\description{
function to filter the hard coded structure required to run the probabilities by the region, country and variants specified in the higher level lists.
}
