% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_by_region_country_and_variant.R
\name{filter_plot_by_region_country_and_variant}
\alias{filter_plot_by_region_country_and_variant}
\title{Filter Plotting Data by Country and Variant}
\usage{
filter_plot_by_region_country_and_variant(
  data,
  COUNTRY,
  VARIANT,
  region_or_country
)
}
\arguments{
\item{data}{data to be filtered}

\item{COUNTRY}{a string containing the name of a country you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}

\item{region_or_country}{a string containing either region or country}
}
\value{
filtered data
}
\description{
filters data to be plotting by country and variant
}
