% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pop_table.R
\name{filter_pop_table}
\alias{filter_pop_table}
\title{filter_pop_table}
\usage{
filter_pop_table(
  population_counts,
  main_config,
  country_configs,
  country_list,
  variant_list
)
}
\arguments{
\item{population_counts}{population counts input data.frame}

\item{main_config}{main config}

\item{country_configs}{a list of country configs}

\item{country_list}{A list of the four UK countries, England, Scotland, Wales and Northern_Ireland}

\item{variant_list}{a list of the variant 1,2,4,5}
}
\description{
filters and cleans the whole uk population counts into DAs and England tables
}
