% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aggregated_samp.R
\name{filter_sample_by_country_region_variant_date}
\alias{filter_sample_by_country_region_variant_date}
\title{filter_sample_by_country_and_region}
\usage{
filter_sample_by_country_region_variant_date(
  country,
  variant,
  aggregated_sample,
  country_configs
)
}
\arguments{
\item{country}{a string containing the name of a country you want to filter by}

\item{variant}{a numeric containing the variant you want to filter by}

\item{aggregated_sample}{cleaned aggregated_sample to be filtered for the country x variant models to be fitted}

\item{country_configs}{a list of country configs}
}
\description{
filters and cleans by country, by dates, by variants and filters days with no returns
}
