% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_prevalence_time_series.R
\name{find_previous_time_series_path}
\alias{find_previous_time_series_path}
\title{Find Previous Time Series Path}
\usage{
find_previous_time_series_path(main_config)
}
\arguments{
\item{main_config}{a list object containing the configuration settings for the system}
}
\value{
a string containing the previous prevelance series filepath
}
\description{
function to find the previous time series file path using the system configuration settings
}
