% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_fit_all_models.R
\name{fit_model}
\alias{fit_model}
\title{fits MRP model}
\usage{
fit_model(data, config)
}
\arguments{
\item{data}{dataframe}

\item{config}{config list}
}
\value{
stanreg object containing the estimated prevalences.
See http://127.0.0.1:35089/help/library/rstanarm/help/stanreg-objects for more info on the returned object.
}
\description{
Takes various parameters and fits a general additive mixed model. The model fits a spline to study day.
}
