% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{format_config_end_dates}
\alias{format_config_end_dates}
\title{format_config_end_dates}
\usage{
format_config_end_dates(config)
}
\arguments{
\item{config}{a list object containing the config}
}
\value{
a list object containing reformatted config with end_date and prev_end_date formatted as date type.
}
\description{
formats the end_date and prev_end_date to date format
}
