% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{format_visit_dates}
\alias{format_visit_dates}
\title{format_visit_dates}
\usage{
format_visit_dates(data)
}
\arguments{
\item{data}{a dataframe containing the data with dates to reformat}
}
\value{
a dataframe with the reformatted visit_dates
}
\description{
function to format visit dates from string (day month year) to date format
}
