% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{get_bucket_list}
\alias{get_bucket_list}
\title{get_bucket_list}
\usage{
get_bucket_list(config)
}
\arguments{
\item{config}{a list object containing the configurations settings}
}
\value{
a list object containing the gcp bucket references for saving outputs
}
\description{
function to retrieve the bucket list from the config
}
