% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_numbers_statement.R
\name{get_country_pop_total}
\alias{get_country_pop_total}
\title{Get Country Population Totals}
\usage{
get_country_pop_total(filtered_pop_tables, country)
}
\arguments{
\item{filtered_pop_tables}{a dataframe containing the population totals by demographic}

\item{country}{a string containing the name of a country you want to filter by}
}
\value{
a value containing the aggregated population for that given country
}
\description{
retrieves a specific countries population totals from the filtered_pop_tables
}
