% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{get_folder_name_prefix}
\alias{get_folder_name_prefix}
\title{create_folder_name_prefix}
\usage{
get_folder_name_prefix(config)
}
\arguments{
\item{config}{a list object containing the main configuration settings for the system}
}
\value{
a string containing prefix for the folder_name
}
\description{
function to create prefix for folder_name from configuration file
}
