% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_lists.R
\name{get_map_lists}
\alias{get_map_lists}
\title{get_map_lists}
\usage{
get_map_lists(countries, country_configs)
}
\arguments{
\item{countries}{a list of countries being run in the pipeline}

\item{country_configs}{a list of country_configs describing the parameter for the countries being run in the pipeline}
}
\value{
list length 2: containing variant list (a list of all variants found for countries) and country a corresponding list of the same length denoting the corresponding country for each variant in variant list.
}
\description{
Uses the configs to generate lists of variants and countries to iterate over.
}
