% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{get_max_date}
\alias{get_max_date}
\title{get_max_date}
\usage{
get_max_date(config, end_date)
}
\arguments{
\item{config}{a list object containing the configurations settings}

\item{end_date}{a string containing the name of a specfic date reference in the config}
}
\value{
a string containing the maximum date from a list in character form
}
\description{
function to retrieve the latest date from a list of dates
}
