% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_formula.R
\name{get_model_formula}
\alias{get_model_formula}
\title{get_model_formula}
\usage{
get_model_formula(
  model,
  country = c("England", "Wales", "Scotland", "Northern Ireland"),
  variant = c("1", "2", "4", "5")
)
}
\arguments{
\item{model}{list of model objects}

\item{country}{character string of country to extract from list}

\item{variant}{character string of variant to extract from list}
}
\description{
for a given stan model return the model formula
}
