% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_divergent.R
\name{get_n_divergent}
\alias{get_n_divergent}
\title{get_n_divergent}
\usage{
get_n_divergent(
  model,
  country = c("England", "Wales", "Scotland", "Northern Ireland"),
  variant = c("1", "2", "4", "5")
)
}
\arguments{
\item{model}{list of model objects}

\item{country}{country to extract from list}

\item{variant}{variant to extract from list}
}
\value{
number of divergent transitions
}
\description{
for a given stan model return the number of divergent transistions
}
