% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_numbers_statement.R
\name{get_population_numbers_statement}
\alias{get_population_numbers_statement}
\title{Get Population Numbers Statement}
\usage{
get_population_numbers_statement(
  filtered_pop_tables,
  prevalence_time_series,
  REGION,
  VARIANT
)
}
\arguments{
\item{filtered_pop_tables}{a dataframe containing the population totals by demographic}

\item{prevalence_time_series}{a dataframe containing the prevalence time series}

\item{REGION}{a string containing the name of the region you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}
}
\value{
a sentence summarising the estimated number of people with covid with credible intervals
}
\description{
create a sentence summarising the estimated number of people with covid-19
}
