% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{get_positives_and_totals_from_raw}
\alias{get_positives_and_totals_from_raw}
\title{get_positives_and_totals_from_raw}
\usage{
get_positives_and_totals_from_raw(data)
}
\arguments{
\item{data}{a dataframe containing the data from which to retrieve the positives and totals}
}
\value{
a dataframe with the positives and totals
}
\description{
function to get positive totals from raw aggregated data
}
