% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_probabilities.R
\name{get_posterior_probabilities}
\alias{get_posterior_probabilities}
\title{get_posterior_probabilities}
\usage{
get_posterior_probabilities(
  models,
  filtered_pop_tables,
  filtered_aggregates,
  country_list,
  variant_list
)
}
\arguments{
\item{models}{list of models produced by fit_models (country x variant)}

\item{filtered_pop_tables}{list of filtered pop tables for the countries}

\item{filtered_aggregates}{list of filtered aggregates i.e. country by variant sample datasets}

\item{country_list}{list of countries}

\item{variant_list}{list of variants}
}
\value{
list  list of posterior probablilities
}
\description{
gets the posterior probabilities from the models
}
