% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_population_percentage_statement.R
\name{get_previous_population_percentage_statement}
\alias{get_previous_population_percentage_statement}
\title{Get Previous Population Percentage Statement}
\usage{
get_previous_population_percentage_statement(
  previous_prevalence_time_series,
  REGION,
  VARIANT
)
}
\arguments{
\item{previous_prevalence_time_series}{a dataframe containing the prevalence time series}

\item{REGION}{a string containing the name of the region you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}
}
\value{
a sentence summarising the estimated number of people with covid with credible intervals
}
\description{
create a sentence summarising the previous population perentage statement
}
