% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_prevalence_time_series.R
\name{get_previous_prevalence_time_series}
\alias{get_previous_prevalence_time_series}
\title{Get Previous Prevalence Time Series}
\usage{
get_previous_prevalence_time_series(main_config)
}
\arguments{
\item{main_config}{a list object containing the configuration settings for the system}
}
\value{
a dataframe containing the previous prevalence time series
}
\description{
retrieve the previous prevalence time series
}
