% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_probabilities.R
\name{get_previous_probabilities}
\alias{get_previous_probabilities}
\title{get_previous_probabilities}
\usage{
get_previous_probabilities(main_config)
}
\arguments{
\item{main_config}{a list object containing the main configuration settings}
}
\value{
a dataframe with the previous weeks probablities
}
\description{
function to fetch previous probabilities from the review bucket
}
