% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{get_reference_dates}
\alias{get_reference_dates}
\title{Get reference dates}
\usage{
get_reference_dates(
  country_configs,
  country = c("England", "Scotland", "Wales", "Northern Ireland")
)
}
\arguments{
\item{country_configs}{a list object containing the country configuration settings}

\item{country}{a string containing the name of the country you want to filter by}
}
\value{
list of reference dates
}
\description{
gets reference dates from country_config for each country
}
