% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{get_rolling_means}
\alias{get_rolling_means}
\title{get_rolling_means}
\usage{
get_rolling_means(data, country_filter, region_or_country, country_configs)
}
\arguments{
\item{data}{a dataframe from which the rolling means will be calculated}

\item{country_filter}{a string containing the country to be filtered}

\item{region_or_country}{a string containing either 'region' or 'country'}

\item{country_configs}{a list object containing the configuration settings}
}
\value{
a data frame with 3-day rolling means
}
\description{
function to calculate 3-day rolling means and append to dataframe
}
