% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{get_sample_and_hh_rolling_means}
\alias{get_sample_and_hh_rolling_means}
\title{get_sample_and_hh_rolling_means}
\usage{
get_sample_and_hh_rolling_means(
  sample_data,
  household_data,
  country_filter,
  country_configs,
  region_or_country
)
}
\arguments{
\item{sample_data}{a dataframe containing the aggregated_sample}

\item{household_data}{a dataframe containing the aggregated_HH the aggregated household datafor}

\item{country_filter}{string containing the name of a country to filter by}

\item{country_configs}{a list object containing the configuration settings}

\item{region_or_country}{string containing 'region' or 'country' levels}
}
\value{
a list containing both the sample_data and the household_data with a 3-day rolling positivity_rate
}
\description{
parent function to get sample and household 3-day rolling means
}
