% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_statement.R
\name{get_summary_statement}
\alias{get_summary_statement}
\title{Get Summary Statement}
\usage{
get_summary_statement(prevalence_time_series, REGION, VARIANT)
}
\arguments{
\item{prevalence_time_series}{a dataframe containing the prevalence time series}

\item{REGION}{a string containing the name of the region you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}
}
\value{
a sentence summarising the latest week/s trend statements.
}
\description{
create a sentence summarising the latest trends
}
