% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{ingest_data}
\alias{ingest_data}
\title{Ingest Data Frame}
\usage{
ingest_data(file_reference, data_run)
}
\arguments{
\item{file_reference}{the partial name of the file you want to read in}

\item{data_run}{a string containing the data run date}
}
\value{
data frame
}
\description{
load csv dataset as dataframe
}
