% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{join_raw_data}
\alias{join_raw_data}
\title{Join Raw Data}
\usage{
join_raw_data(cleaned_raw_data)
}
\arguments{
\item{cleaned_raw_data}{a list object containing clean_sample and clean_hh dataframes}
}
\value{
a dataframe with the particpant and household data combined
}
\description{
a function to join the cleaned household and participant data
into a single data frame
}
