% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{join_raw_with_prevalence}
\alias{join_raw_with_prevalence}
\title{Join Raw with Prevalence}
\usage{
join_raw_with_prevalence(cleaned_raw_data, prevalence_time_series)
}
\arguments{
\item{cleaned_raw_data}{a dataframe containing the cleaned and joined particpant and
household raw data}

\item{prevalence_time_series}{a dataframe containing the modelled prevalence estimates}
}
\value{
a dataframe with the raw and modelled data combined together
}
\description{
a function to join the cleaned and joined particpant and household data
with the modelled prevalence time series.
}
