% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{lag_twice_and_sum}
\alias{lag_twice_and_sum}
\title{lag_twice_and_sum}
\usage{
lag_twice_and_sum(vector)
}
\arguments{
\item{vector}{a vector of data for which to lag twice and sum}
}
\value{
a new vector which has been lagged twice and summed
}
\description{
function to lag a vector two times and sum the values
}
