% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_country_configs.R
\name{load_country_configs}
\alias{load_country_configs}
\title{load the country configuration files}
\usage{
load_country_configs(countries, main_config)
}
\arguments{
\item{countries}{list of countries}

\item{main_config}{overall pipeline config settings}
}
\value{
list of country configurations
}
\description{
loads pipeline configurations from country yaml files
}
