% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_divergent_table.R
\name{make_divergent_table}
\alias{make_divergent_table}
\title{Make Divergent Table}
\usage{
make_divergent_table(country_configs, country, models, main_config)
}
\arguments{
\item{country_configs}{a list object containing the configuration settings}

\item{country}{a string containing the name of a country you want to filter by}

\item{models}{a list object containing the model}

\item{main_config}{a list object containing the main configuration settings for the system}
}
\value{
a kable table
}
\description{
This function creates a kable table listing the number of divergent transitions.
}
