% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_vs_prevalence_table.R
\name{make_trimmed_raw_and_prevalence}
\alias{make_trimmed_raw_and_prevalence}
\title{Make Trimmed Raw and Prevalence}
\usage{
make_trimmed_raw_and_prevalence(raw_and_prevalence, region_or_country)
}
\arguments{
\item{raw_and_prevalence}{data frame}

\item{region_or_country}{character either "region" or "country"}
}
\value{
a kable table
}
\description{
Takes in a raw and prevalence table and formats it using the kable package.
}
