% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pop_table.R
\name{merge_boost_regions}
\alias{merge_boost_regions}
\title{merge_boost_regions}
\usage{
merge_boost_regions(pop_table)
}
\arguments{
\item{pop_table}{data frame}
}
\value{
data frame
}
\description{
merge _0, _1 and _2 region names, effectively this removes boost regions from the data
}
