% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_fit_all_models.R
\name{par_Map}
\alias{par_Map}
\title{par_Map}
\usage{
par_Map(f, cores = 16, ...)
}
\arguments{
\item{f}{a function to parallelise over}

\item{cores}{number of cores (defaults to 16)}

\item{...}{further arguments to f}
}
\value{
NA (depends on function)
}
\description{
parrallelise over a function
}
