% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{plot_positivity_and_raw}
\alias{plot_positivity_and_raw}
\title{plot_positivity_and_raw}
\usage{
plot_positivity_and_raw(
  prevalence_time_series,
  sample_rolling_means,
  household_rolling_means,
  region_or_country
)
}
\arguments{
\item{prevalence_time_series}{MRP modelled_data series}

\item{sample_rolling_means}{data frame with 3-day rolling test positivity}

\item{household_rolling_means}{data frame with 3-day rolling household positivity}

\item{region_or_country}{whether to plot by region}
}
\value{
overlay plot with mrp estimates and raw data, and horizontally aligned regional plots
}
\description{
generates a plot with modelled positivity, household rolling positivity and test rolling positivity
}
