% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{plot_raw_vs_modelled}
\alias{plot_raw_vs_modelled}
\title{Plot Raw vs Modelled}
\usage{
plot_raw_vs_modelled(
  sample_data,
  household_data,
  prevalence_time_series,
  COUNTRY,
  country_configs,
  region_or_country
)
}
\arguments{
\item{sample_data}{a dataframe containing the aggregated_sample}

\item{household_data}{a dataframe containing the aggregated_HH the aggregated household datafor}

\item{prevalence_time_series}{a dataframe containing the prevalence time series}

\item{COUNTRY}{string containing the name of a country to filter by}

\item{country_configs}{a list object containing the configuration settings}

\item{region_or_country}{string containing 'region' or 'country' levels}
}
\description{
plot the raw household and sample data against the modelled data
}
