% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_vs_modelled.R
\name{positivity_rolling_mean}
\alias{positivity_rolling_mean}
\title{positivity_rolling_mean}
\usage{
positivity_rolling_mean(data)
}
\arguments{
\item{data}{a dataframe containing the raw data}
}
\value{
a data frame with 3-day rolling means appended
}
\description{
function to caluclate the 3-day rolling mean for raw positivity estimates
}
