% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify_posterior_probs.R
\name{post_stratify_draws}
\alias{post_stratify_draws}
\title{post stratify draws}
\usage{
post_stratify_draws(post_strat_table, posterior_prob, n_days_to_model)
}
\arguments{
\item{post_strat_table}{population table with counts for every combination of conditions from each factor in model, with Ns.
MUST have the same conditions as the table used in rstanarm::postrior_epred / rstanarm::posterior_linpred. if
no argumant is passed to n_days_to_model, the table much be replicated for every day of the model
and day number must be indicated by study_day variable.}

\item{posterior_prob}{matrix of posterior draws outputted by posterior_epred / posterior_linpred.
n rows = number of draws sampled. n columns = the product of the number
of levels in each predictor x n days modelled. each column of posterior_probs
corresponds to each row in the post-stratification table assigned to new_data
during extraction with posterior_linpred}

\item{n_days_to_model}{optional argument (but not recommended). used if the user wants to use a
poststrat table that has not been replicated for every day of the model.
poststrat table must still have a study_day column.
best to use the same post_strat table that was used to extract data from model
fit object which would hhave been replicated, rather than this option.}
}
\value{
Dataframe containing posterior draws post-stratified according to population counts in each group combinations
n rows = number of draws samples in model. n columns = number of days modelled.

post-stratified draws
}
\description{
post-stratifies posterior draws that have been extracted from model fit using rstanarm::posterior_epred
or posterior_linpred and data = post_strat_table. For each day, draws are matrix multiplied
with the proportion that N corresponding to the factor level combination makes up.
}
