% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify_posterior_probs.R
\name{post_stratify_draws_by_region}
\alias{post_stratify_draws_by_region}
\title{post stratify draws by region}
\usage{
post_stratify_draws_by_region(posteriors_list, post_strat_table)
}
\arguments{
\item{posteriors_list}{list of matrices, returned from extract_posterior_probabilities}

\item{post_strat_table}{population counts for each group - see post_stratify_draws}
}
\value{
list of post-stratified draws
}
\description{
loop through all regions and post stratify draws
}
