% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify_posterior_probs.R
\name{post_stratify_posterior_probs}
\alias{post_stratify_posterior_probs}
\title{post_stratify_posterior_probs}
\usage{
post_stratify_posterior_probs(
  filtered_pop_tables,
  posterior_probabilities,
  country_list,
  variant_list
)
}
\arguments{
\item{filtered_pop_tables}{list of filtered pop tables for the countries}

\item{posterior_probabilities}{list of posterior probablilities to apply post-stratification to}

\item{country_list}{list of countries}

\item{variant_list}{list of variants}
}
\value{
list a list of poststratified draws (country x variant)
}
\description{
post-stratified draws from the posterior probabilities
}
