% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{prepare_main_config}
\alias{prepare_main_config}
\title{prepare_main_config}
\usage{
prepare_main_config(config_filename)
}
\arguments{
\item{config_filename}{a string containing the filename of the main config}
}
\value{
a formatted config which has been checked for user errors
}
\description{
formats the end_date and prev_end_date to date format
}
