% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{prepare_sample_and_household_for_joining}
\alias{prepare_sample_and_household_for_joining}
\title{Prepare Sample and Household Data for Joining}
\usage{
prepare_sample_and_household_for_joining(aggregated_sample, aggregated_HH)
}
\arguments{
\item{aggregated_sample}{a dataframe containing the raw participant sample}

\item{aggregated_HH}{a dataframe containing the raw household sample}
}
\value{
a list object containing the cleaned participant and household samples
}
\description{
carry out prepatory cleaning and formatting of the aggregated sample and
aggregated household files so that they are compatible for joining
}
