% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/probabilities_compared_to_previously_published.R
\name{probabilities_compared_to_previously_published}
\alias{probabilities_compared_to_previously_published}
\title{probabilities_compared_to_previously_published}
\usage{
probabilities_compared_to_previously_published(
  probabilities,
  previous_probabilities
)
}
\arguments{
\item{probabilities}{a dataframe which you want to reformat the variant column of}

\item{previous_probabilities}{a dataframe containing the previous probabilities}
}
\value{
a dataframe with the probabilities
}
\description{
function to compare current and previous probabilities and prepare table for use in qa report
}
