% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{probability_of_at_least_x_percent_increase}
\alias{probability_of_at_least_x_percent_increase}
\title{probability_of_at_least_x_percent_increase}
\usage{
probability_of_at_least_x_percent_increase(data, country_configs, country)
}
\arguments{
\item{data}{a dataframe containing reference_day and week_minus_1 and week_minus_2 columns}

\item{country_configs}{a list object containing the configuration settings for the country}

\item{country}{a string containing the name of the country you want to filter by}
}
\value{
a dataframe with the added columns
}
\description{
function to calculate the probability of being x% above previous weeks estimates.
i.e what percentage of the reference_day draws are 15 or more percent above the comparison dates draws on average
}
