% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{probability_of_increase}
\alias{probability_of_increase}
\title{probability_of_increase}
\usage{
probability_of_increase(data)
}
\arguments{
\item{data}{a dataframe containing reference_day and week_minus_1 and week_minus_2 columns}
}
\value{
a dataframe with the added columns
}
\description{
function to calculate the probability of being above 1 and 2 weeks
}
