% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{probability_reference_at_least_x_percent_greater_than_comparison}
\alias{probability_reference_at_least_x_percent_greater_than_comparison}
\title{probability_reference_at_least_x_percent_greater_than_comparison}
\usage{
probability_reference_at_least_x_percent_greater_than_comparison(
  reference,
  comparison,
  percent_change_threshold
)
}
\arguments{
\item{reference}{a vector of draws you want to use as your reference group in comparisons}

\item{comparison}{a vector of draws you want to use as your comparison group}

\item{percent_change_threshold}{a numeric value between 0-100 which to use as the threshold to determine significant change.}
}
\value{
a dataframe with the added columns
}
\description{
function to calculate the probability that a reference group of draws is at least x percent (decided by the percent_change_threshold) greater than the comparison group.
}
