% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report.R
\name{qa_report}
\alias{qa_report}
\title{QA Report}
\usage{
qa_report(
  local_location,
  bucket,
  main_config,
  country_configs,
  prevalence_time_series,
  previous_prevalence_time_series,
  reference_day_vs_weeks_before_probabilities,
  cleaned_aggregated_sample,
  aggregated_HH,
  aggregated_sample,
  post_stratified_draws
)
}
\arguments{
\item{local_location}{the local file path where you want to save the
qa_report}

\item{bucket}{the reference for the gcp bucket where you want to save this
to}

\item{main_config}{a list object containing the configuration settings}

\item{country_configs}{a list object containing the country specific
configuration settings}

\item{prevalence_time_series}{the main prevalence time series}

\item{previous_prevalence_time_series}{the previous prevalence time series}

\item{reference_day_vs_weeks_before_probabilities}{the week on week
probability comparisons}

\item{cleaned_aggregated_sample}{the cleaned aggregated sample}

\item{aggregated_HH}{the raw aggregated household data}

\item{aggregated_sample}{the raw aggregated sample data}

\item{post_stratified_draws}{the post-stratified draws from the model}
}
\value{
invisible, creates a qa_report file in the given location
}
\description{
Creates quality assurance report from the latest run of data
}
