% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_population_counts.R
\name{recode_age_groups}
\alias{recode_age_groups}
\title{recode age groups}
\usage{
recode_age_groups(data)
}
\arguments{
\item{data}{data frame}
}
\value{
data frame
}
\description{
recode age groups ("_" to "-") and rename column
}
