% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_population_counts.R
\name{recode_pop_region_codes}
\alias{recode_pop_region_codes}
\title{recode_pop_region_codes}
\usage{
recode_pop_region_codes(data)
}
\arguments{
\item{data}{a dataframe containing a gor9d column}
}
\value{
a dataframe containing readable region names, and the gor9d column
}
\description{
function to derive the readable region names from the gor_code column
}
