% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_region.R
\name{recode_region_names}
\alias{recode_region_names}
\title{recode_region_names}
\usage{
recode_region_names(data, column)
}
\arguments{
\item{data}{a dataframe containing the region codes for renaming}

\item{column}{a column name which contains the region codes for renaming}
}
\value{
a vector with the recoded names in a new column called 'region'
}
\description{
function to recode the region names from alpha-numeric code to readable region names
}
