% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{recode_variant_names}
\alias{recode_variant_names}
\title{recode_variant_names}
\usage{
recode_variant_names(data)
}
\arguments{
\item{data}{a probabilities dataframe which you want to reformat the variant column of}
}
\value{
a dataframe with the recoded variant column
}
\description{
function to recode variant names from numbers to ct reference strings
}
