% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_region.R
\name{region_to_factor}
\alias{region_to_factor}
\title{region_to_factor}
\usage{
region_to_factor(data)
}
\arguments{
\item{data}{a dataframe containing the region codes for renaming}
}
\value{
a vector with the recoded names
}
\description{
function to convert a region column into an ordered factor
}
