% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variant_overlay.R
\name{relabel_variants}
\alias{relabel_variants}
\title{Relabel Variants}
\usage{
relabel_variants(prevalence_time_series, main_config)
}
\arguments{
\item{prevalence_time_series}{MRP modelled_data series}

\item{main_config}{a list object containing the configuration settings for the system}
}
\value{
a dataframe with the relabeled variants
}
\description{
re-label variant column with the replacements from the main_config
}
