% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_aggregated.R
\name{remove_ethnicity}
\alias{remove_ethnicity}
\title{remove_ethnicity}
\usage{
remove_ethnicity(data)
}
\arguments{
\item{data}{dataframe containing data to aggregate}
}
\value{
a dataframe with ethnicity removed.
}
\description{
function to aggregate up to remove ethnicity breakdown
}
