% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_main_config.R
\name{run_config_checks}
\alias{run_config_checks}
\title{run_config_checks}
\usage{
run_config_checks(config)
}
\arguments{
\item{config}{- a list object containing the configurations for the system}
}
\value{
warning messages or errors if configuration is set incorrectly
}
\description{
checking config is compatible with model
}
