% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_all_outputs}
\alias{save_all_outputs}
\title{save_all_outputs}
\usage{
save_all_outputs(objects, file_ids, file_types, config)
}
\arguments{
\item{objects}{a list object containing the objects you want to save}

\item{file_ids}{a list object containing the identifying name components to fit within the naming pattern}

\item{file_types}{a list containing strings the file extension without the '.' e.g. 'rdata' or 'csv'}

\item{config}{a list object containing the configurations settings}
}
\value{
nothing
}
\description{
function to save a set of outputs to the gcp buckets
}
