% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_output}
\alias{save_output}
\title{save_output}
\usage{
save_output(object, file_id, file_type, config)
}
\arguments{
\item{object}{a data object containing the thing you want to save}

\item{file_id}{a string containing the identifying characteristics of the object}

\item{file_type}{a list containing strings the file extension without the '.' e.g. 'rdata' or 'csv'}

\item{config}{a list object containing the configurations settings}
}
\value{
nothing
}
\description{
function to save outputs to a given bucket
}
