% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_population_counts.R
\name{set_pop_table_factors}
\alias{set_pop_table_factors}
\title{set_pop_table_factors}
\usage{
set_pop_table_factors(data)
}
\arguments{
\item{data}{a dataframe containing region and age_group columns}
}
\value{
data frame with factor levels set
}
\description{
function to set the factor levels for region and age groups
}
