% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Spaghetti Plot}
\usage{
spaghetti_plot(
  prevalence_time_series,
  post_stratified_draws,
  COUNTRY,
  VARIANT,
  region_or_country,
  main_config
)
}
\arguments{
\item{prevalence_time_series}{a dataframe containing the prevalence time series}

\item{post_stratified_draws}{a dataframe of poststratified draws}

\item{COUNTRY}{a string containing the name of a country you want to filter by}

\item{VARIANT}{a string containing the name of the variant you want to filter by}

\item{region_or_country}{specify "country" for one ggplot object of data from an individial country, specify "region" for ggplot object of England regions}

\item{main_config}{list of configuration settings}
}
\value{
a ggplot object containing the spaghetti plot
}
\description{
creates a plot showing how the individual sample draws come together to create the mean and credible intervals
}
