% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_aggregated_HH}
\alias{synthetic_aggregated_HH}
\title{Synthetic Aggregated Housholds}
\format{
An object of class \code{data.frame} with 1510 rows and 11 columns.
}
\source{
created by data-raw/make_synthetic_aggregated_hh.R
}
\usage{
synthetic_aggregated_HH
}
\description{
A synthetic  aggregated household version of the data which is after it is first loaded in
}
\seealso{
Other Synthetic Data: 
\code{\link{synthetic_aggregated_sample}},
\code{\link{synthetic_cleaned_aggregated_sample}},
\code{\link{synthetic_cleaned_population_counts}},
\code{\link{synthetic_filtered_aggregates}},
\code{\link{synthetic_filtered_pop_tables}},
\code{\link{synthetic_overview}},
\code{\link{synthetic_population_counts}},
\code{\link{synthetic_prevalence_time_series}},
\code{\link{synthetic_previous_prevalence_time_series}}
}
\concept{Synthetic Data}
\keyword{datasets}
