% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_cleaned_aggregated_sample}
\alias{synthetic_cleaned_aggregated_sample}
\title{Synthetic Cleaned Aggregated Sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14026 rows and 17 columns.
}
\source{
created by data-raw/make_synthetic_data.R
}
\usage{
synthetic_cleaned_aggregated_sample
}
\description{
A synthetic dataset used for examples and/or testing. Represents what aggregated sample should look like after ingest and cleaning applied. This synthetic data represents the England and Scotland data.
}
\seealso{
Other Synthetic Data: 
\code{\link{synthetic_aggregated_HH}},
\code{\link{synthetic_aggregated_sample}},
\code{\link{synthetic_cleaned_population_counts}},
\code{\link{synthetic_filtered_aggregates}},
\code{\link{synthetic_filtered_pop_tables}},
\code{\link{synthetic_overview}},
\code{\link{synthetic_population_counts}},
\code{\link{synthetic_prevalence_time_series}},
\code{\link{synthetic_previous_prevalence_time_series}}
}
\concept{Synthetic Data}
\keyword{datasets}
