% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_cleaned_population_counts}
\alias{synthetic_cleaned_population_counts}
\title{Synthetic Cleaned Population Counts}
\format{
An object of class \code{data.frame} with 168 rows and 6 columns.
}
\source{
created by data-raw/make_synthetic_data.R
}
\usage{
synthetic_cleaned_population_counts
}
\description{
A synthetic dataset used for examples and/or testing. Represents what population counts should look like after initial cleaning. This synthetic data represents the England and Scotland data.
}
\seealso{
Other Synthetic Data: 
\code{\link{synthetic_aggregated_HH}},
\code{\link{synthetic_aggregated_sample}},
\code{\link{synthetic_cleaned_aggregated_sample}},
\code{\link{synthetic_filtered_aggregates}},
\code{\link{synthetic_filtered_pop_tables}},
\code{\link{synthetic_overview}},
\code{\link{synthetic_population_counts}},
\code{\link{synthetic_prevalence_time_series}},
\code{\link{synthetic_previous_prevalence_time_series}}
}
\concept{Synthetic Data}
\keyword{datasets}
