% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_filtered_aggregates}
\alias{synthetic_filtered_aggregates}
\title{Synthetic Filtered Aggregates}
\format{
An object of class \code{list} of length 8.
}
\source{
created by data-raw/make_synthetic_data.R
}
\usage{
synthetic_filtered_aggregates
}
\description{
A synthetic dataset used for examples and/or testing. Represents what aggregated sample should look like after ingest, cleaning and filtering applied. This synthetic data represents the England and Scotland data.
}
\seealso{
Other Synthetic Data: 
\code{\link{synthetic_aggregated_HH}},
\code{\link{synthetic_aggregated_sample}},
\code{\link{synthetic_cleaned_aggregated_sample}},
\code{\link{synthetic_cleaned_population_counts}},
\code{\link{synthetic_filtered_pop_tables}},
\code{\link{synthetic_overview}},
\code{\link{synthetic_population_counts}},
\code{\link{synthetic_prevalence_time_series}},
\code{\link{synthetic_previous_prevalence_time_series}}
}
\concept{Synthetic Data}
\keyword{datasets}
