% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{synthetic_overview}
\alias{synthetic_overview}
\title{Synthetic Data Overview}
\description{
Synthetic data is generated using the Synthpop library to synthesise population and sample columns so data is non-disclosive but attempts to capture trends present in the original dataset. Columns have been synthesised using the CART method. Guassian kernel density smoothing is applied after synthesis.
}
\seealso{
Other Synthetic Data: 
\code{\link{synthetic_aggregated_HH}},
\code{\link{synthetic_aggregated_sample}},
\code{\link{synthetic_cleaned_aggregated_sample}},
\code{\link{synthetic_cleaned_population_counts}},
\code{\link{synthetic_filtered_aggregates}},
\code{\link{synthetic_filtered_pop_tables}},
\code{\link{synthetic_population_counts}},
\code{\link{synthetic_prevalence_time_series}},
\code{\link{synthetic_previous_prevalence_time_series}}
}
\concept{Synthetic Data}
