% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc_timer.R
\name{tictoc_timer}
\alias{tictoc_timer}
\title{tictoc_timer}
\usage{
tictoc_timer(n_toc = 8, digits = 2, log)
}
\arguments{
\item{n_toc}{the toc number you wish to measure the time until}

\item{digits}{n_digits of precision}

\item{log}{an unformatted tictoc log}
}
\description{
records the amount of time in minutes from the first tic of the monitored R code to a given toc in the log
}
