% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{transform_draws_to_df}
\alias{transform_draws_to_df}
\title{Transform Draws to a Data Frame}
\usage{
transform_draws_to_df(
  COUNTRY,
  VARIANT,
  REGION,
  post_stratified_draws,
  main_config
)
}
\arguments{
\item{COUNTRY}{a string containing the name of the country which the draws relate to}

\item{VARIANT}{a string containing the name of the variant which the draws relate to}

\item{REGION}{a string contianing the name of the region which the draws relate to}

\item{post_stratified_draws}{a list object containing the complete set of post-stratified draws matricies}

\item{main_config}{a list object containing the main configuration settings for the system}
}
\value{
a dataframe for the specific region, country and variant post-stratified draws
}
\description{
selects and transforms a post-stratifed draws matrix from the wider list into a dataframe with date, region, country and variant columns
}
