% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raw_counts_with_prevalence.R
\name{transform_gor_name_to_full_name}
\alias{transform_gor_name_to_full_name}
\title{Transform gor_name to Full Geography Area Name}
\usage{
transform_gor_name_to_full_name(data)
}
\arguments{
\item{data}{a data frame containing the gor_name column}
}
\value{
a dataframe with the NUTS1 regions and UK countries
}
\description{
function to format gor_names to full names for region and country
}
