% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_fit_all_models.R
\name{try_catch_fit_all_models}
\alias{try_catch_fit_all_models}
\title{try_catch_fit_all_models}
\usage{
try_catch_fit_all_models(
  country_configs,
  country_list,
  variant_list,
  filtered_aggregates,
  cores = 16
)
}
\arguments{
\item{country_configs}{a list of country configs}

\item{country_list}{A list of the four UK countries, England, Scotland, Wales and Northern_Ireland}

\item{variant_list}{a list of the variants 1, 2, 4, 5}

\item{filtered_aggregates}{list of filtered aggregates i.e. country by variant sample datasets}

\item{cores}{the number of cores used by par_Map (defaults to 16)}
}
\description{
fits GAMM for each country x variant combination provided
}
