% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_aggregated_sample_variant_positives.R
\name{validate_aggregated_sample_variant_positives}
\alias{validate_aggregated_sample_variant_positives}
\title{validate_aggregated_sample_variant_positives}
\usage{
validate_aggregated_sample_variant_positives(data)
}
\arguments{
\item{data}{the aggregated sample data}
}
\description{
basic checks on the input data and stops the script if the input data is incorrect.
}
