% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities_comparing_dates.R
\name{wrangle_probabilities_to_shape}
\alias{wrangle_probabilities_to_shape}
\title{wrangle_probabilities_to_shape}
\usage{
wrangle_probabilities_to_shape(data)
}
\arguments{
\item{data}{a probabilities dataframe which you want to get into shape}
}
\value{
a reformatted dataframe which has removed duplication and reshaped for interpretation.
}
\description{
function to format probabilities data into correct shape
}
